Refactor UsedPartialKeysList to track a single key
