+++ src/rpc/cluster/submit_decryption_key.rs
@@ -47,26 +47,17 @@ impl RpcParameter<AppState> for SubmitDecryptionKey {
             self.payload.timestamp
         );
 
-        // Validation logic - in this example, only checking if Solver is registered in the cluster
-        // if !KeyGeneratorList::get()?.is_key_generator_in_cluster(&sender_address) {
-        //     return Err(RpcError::InvalidParams(format!(
-        //         "Address {} is not a registered key generator",
-        //         sender_address.as_hex_string()
-        //     )));
-        // }
-
         // Store decryption key
-        let _decryption_key = DecryptionKey::new(self.payload.decryption_key.clone());
-        // decryption_key.put(self.payload.key_id)?;
+        let decryption_key = DecryptionKey::new(self.payload.decryption_key.clone());
+        decryption_key.put(self.payload.session_id)?;
 
-        // TODO: Add broadcast logic via ack_decryption_key at this point
-        // This will be implemented in a separate function
+        // TODO: Add broadcast_decryption_key_ack
 
         info!(
             "[{}] Complete to get decryption key - key_id: {:?} / decryption key: {:?}",
             _context.config().address().to_short(),
             self.payload.session_id,
-            _decryption_key
+            decryption_key
         );
 
         Ok(DecryptionKeyResponse { success: true })
