+++ src/rpc/external/get_latest_key_id.rs
@@ -1,25 +1,25 @@
 use crate::rpc::prelude::*;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct GetLatestKeyId {}
+pub struct GetLatestSessionId {}
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct GetLatestKeyIdResponse {
+pub struct GetLatestSessionIdResponse {
     pub latest_session_id: SessionId,
 }
 
-impl RpcParameter<AppState> for GetLatestKeyId {
-    type Response = GetLatestKeyIdResponse;
+impl RpcParameter<AppState> for GetLatestSessionId {
+    type Response = GetLatestSessionIdResponse;
 
     fn method() -> &'static str {
-        "get_latest_key_id"
+        "get_latest_session_id"
     }
 
     async fn handler(self, _context: AppState) -> Result<Self::Response, RpcError> {
         let session_id = SessionId::get()?;
 
         // loop {
-        return Ok(GetLatestKeyIdResponse {
+        return Ok(GetLatestSessionIdResponse {
             latest_session_id: session_id,
         });
         // }
