+++ src/rpc/internal/add_key_generator.rs
@@ -31,13 +31,6 @@ impl RpcParameter<AppState> for AddKeyGenerator {
             self.message.external_rpc_url
         );
 
-        // TODO: Uncomment this code
-        // self.signature.verify_signature(
-        //     serialize_to_bincode(&self.message)?.as_slice(),
-        //     context.config().radius_foundation_address().as_slice(),
-        //     context.config().chain_type().clone(),
-        // )?;
-
         let key_generator = KeyGenerator::new(
             self.message.address.clone(),
             self.message.cluster_rpc_url.clone(),
@@ -60,7 +53,7 @@ impl RpcParameter<AppState> for AddKeyGenerator {
 }
 
 pub fn sync_key_generator(add_key_generator: AddKeyGenerator) {
-    let other_key_generator_rpc_url_list = KeyGeneratorList::get()
+    let key_generator_rpc_url_list = KeyGeneratorList::get()
         .unwrap()
         .get_all_key_generator_rpc_url_list();
 
@@ -69,13 +62,13 @@ pub fn sync_key_generator(add_key_generator: AddKeyGenerator) {
             "Sync distributed key generation - address: {:?} / cluster_rpc_url: {:?} / rpc_client_count: {:?}",
             add_key_generator.message.address.as_hex_string(),
             add_key_generator.message.cluster_rpc_url,
-            other_key_generator_rpc_url_list.len()
+            key_generator_rpc_url_list.len()
         );
 
         let rpc_client = RpcClient::new().unwrap();
         rpc_client
             .multicast(
-                other_key_generator_rpc_url_list,
+                key_generator_rpc_url_list,
                 SyncKeyGenerator::method(),
                 &add_key_generator,
                 Id::Null,
